+++ scripts/batch_customers.php
@@ -755,8 +755,12 @@ if (! $nboferrors) {
 
 	if ($action == 'backup' || $action == 'backupdelete' ||$action == 'backuprsync' || $action == 'backupdatabase' || $action == 'backuptest' || $action == 'backuptestrsync' || $action == 'backuptestdatabase') {
 		if (empty($instancefilter)) {
-			$from = $conf->global->SELLYOURSAAS_NOREPLY_EMAIL;
-			$to = $conf->global->SELLYOURSAAS_SUPERVISION_EMAIL;
+			$from = $emailfrom;
+			$to = $emailsupervision;
+			// Force to use local sending (MAIN_MAIL_SENDMODE is the one of the master server. It may be to an external SMTP server not allowed to the deployment server)
+			$conf->global->MAIN_MAIL_SENDMODE = 'mail';
+			$conf->global->MAIN_MAIL_SMTP_SERVER = 'localhost';
+
 			$msg = 'Backup done without errors on '.gethostname().' by '.$script_file." ".(empty($argv[1]) ? '' : $argv[1])." ".(empty($argv[2]) ? '' : $argv[2])." (finished at ".dol_print_date(dol_now('gmt'), "%Y%m%d-%H%M%S", 'gmt').")\n\n".$out;
 
 			$sellyoursaasname = $conf->global->SELLYOURSAAS_NAME;                 // exemple 'DoliCloud'
@@ -782,9 +786,13 @@ if (! $nboferrors) {
 
 	if ($action == 'backup' || $action == 'backupdelete' ||$action == 'backuprsync' || $action == 'backupdatabase' || $action == 'backuptest' || $action == 'backuptestrsync' || $action == 'backuptestdatabase') {
 		if (empty($instancefilter)) {
-			$from = $conf->global->SELLYOURSAAS_NOREPLY_EMAIL;
-			$to = $conf->global->SELLYOURSAAS_SUPERVISION_EMAIL;
-			// Supervision tools are generic for all domain. No way to target a specific supervision email.
+			$from = $emailfrom;
+			$to = $emailsupervision;
+			// Force to use local sending (MAIN_MAIL_SENDMODE is the one of the master server. It may be to an external SMTP server not allowed to the deployment server)
+			$conf->global->MAIN_MAIL_SENDMODE = 'mail';
+			$conf->global->MAIN_MAIL_SMTP_SERVER = '';
+
+			// Supervision tools are generic for all domains. No way to target a specific supervision email.
 
 			$msg = 'Error in '.$script_file." ".$argv[1]." ".$argv[2]." (finished at ".dol_print_date(dol_now('gmt'), "%Y%m%d-%H%M%S", 'gmt').")\n\n".$out;