+++ scripts/restore_instance.php
@@ -533,6 +533,9 @@ if (empty($return_var) && empty($return_varmysql)) {
 	if ($mode == 'confirm') {
 		$from = $emailfrom;
 		$to = $emailsupervision;
+		// Force to use local sending (MAIN_MAIL_SENDMODE is the one of the master server. It may be to an external SMTP server not allowed to the deployment server)
+		$conf->global->MAIN_MAIL_SENDMODE = 'mail';
+		$conf->global->MAIN_MAIL_SMTP_SERVER = '';
 
 		// Supervision tools are generic for all domain. No way to target a specific supervision email.